+++ features/sharding/core/src/main/java/org/apache/shardingsphere/sharding/route/engine/type/standard/ShardingStandardRouteEngine.java
@@ -260,9 +260,9 @@ public final class ShardingStandardRouteEngine implements ShardingRouteEngine {
             return shardingTable.getActualDataSourceNames();
         }
         Collection<String> result = databaseShardingStrategy.doSharding(shardingTable.getActualDataSourceNames(), databaseShardingValues, shardingTable.getDataSourceDataNode(), props);
-        ShardingSpherePreconditions.checkNotEmpty(result, NoShardingDatabaseRouteInfoException::new);
+        ShardingSpherePreconditions.checkNotEmpty(result, () -> new NoShardingDatabaseRouteInfoException(shardingTable.getActualDataSourceNames(), databaseShardingValues));
         ShardingSpherePreconditions.checkState(shardingTable.getActualDataSourceNames().containsAll(result),
-                () -> new MismatchedShardingDataSourceRouteInfoException(result, shardingTable.getActualDataSourceNames()));
+                () -> new MismatchedShardingDataSourceRouteInfoException(result, shardingTable.getActualDataSourceNames(), databaseShardingValues));
         return result;
     }